flatMapが利用できないため代替
